requirements.txt
pytest
allure-pytest
appium-python-client
opencv-python
pytesseract
python-dotenv
